backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.10
python-dotenv==1.0.0
Pillow==10.445.0
qrcode[pil]==7.4.2
